application/requirement.txt
https://github.com/django-nonrel/django/zipball/nonrel-1.6
https://github.com/django-nonrel/djangotoolbox/zipball/master
https://github.com/django-nonrel/mongodb-engine/zipball/master
django-cors-headers
djangorestframework
twilio
